backend/requirements.txt
# Flask and extensions
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0

# MongoDB
pymongo==4.6.1
python-dotenv==1.0.0

# Password hashing
bcrypt==4.1.2

# OpenAI (default AI provider)
openai>=1.12.0

# Validation
email-validator==2.1.0

# Production server
gunicorn==21.2.0
